my_fscanf
my_fscanf

+++ operators.c
@@ -1,16 +1,16 @@
 #include "operators.h"
 
-double add(int op1, int op2) {
+int add(int op1, int op2) {
 	return op1+op2;
 }
-double minus(int op1, int op2) {
+int minus(int op1, int op2) {
 	return op1-op2;
 }
-double mul(int op1, int op2) {
+int mul(int op1, int op2) {
 	return op1*op2;
 }
 
-double div(int op1, int op2) {
-	return (double)op1%(double)op2;
+int div(int op1, int op2) {
+	return op1%op2;
 }
